refactor!: added new crate co-circom-snarks
BREAKING CHANGE: moved common code for PLONK and
Groth16 into separate crate. Most notably the
SharedWitness and SharedInput
