+++ co-circom-snarks/Cargo.toml
@@ -1,5 +1,6 @@
+
 [package]
-name = "collaborative-plonk"
+name = "co-circom-snarks"
 version = "0.1.0"
 publish.workspace = true
 authors.workspace = true
@@ -12,20 +13,13 @@ rust-version.workspace = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+rand = { workspace = true }
 ark-ec = { workspace = true }
 ark-ff = { workspace = true }
-ark-poly = { workspace = true }
 ark-serialize = { workspace = true }
 circom-types = { version = "0.3.0", path = "../circom-types" }
-co-circom-snarks = { version = "0.1.0", path = "../co-circom-snarks" }
 eyre = { workspace = true }
-itertools = { workspace = true }
 mpc-core = { version = "0.2.1", path = "../mpc-core" }
 num-traits = { workspace = true }
-sha3 = { workspace = true }
-thiserror = { workspace = true }
-
+serde = { workspace = true }
 [dev-dependencies]
-ark-bn254 = { workspace = true }
-num-traits = { workspace = true }
-serde_json = { workspace = true }
