+++ collaborative-circom/Cargo.toml
@@ -13,32 +13,33 @@ rust-version.workspace = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-ark-bn254.workspace = true
 ark-bls12-381.workspace = true
+ark-bn254.workspace = true
 ark-ec.workspace = true
 ark-ff.workspace = true
 bincode.workspace = true
-circom-types = { version = "0.3.0", path = "../circom-types" }
 circom-mpc-compiler = { version = "0.3.0", path = "../circom-mpc-compiler" }
 circom-mpc-vm = { version = "0.2.0", path = "../circom-mpc-vm" }
+circom-types = { version = "0.3.0", path = "../circom-types" }
 clap.workspace = true
-collaborative-plonk = { version = "0.1.0", path = "../collaborative-plonk" }
+co-circom-snarks = { version = "0.1.0", path = "../co-circom-snarks" }
 collaborative-groth16 = { version = "0.2.1", path = "../collaborative-groth16", features = [
     "verifier",
 ] }
+collaborative-plonk = { version = "0.1.0", path = "../collaborative-plonk" }
 color-eyre.workspace = true
+figment.workspace = true
 mpc-core = { version = "0.2.1", path = "../mpc-core" }
 mpc-net.workspace = true
 num-bigint.workspace = true
-rand.workspace = true
 num-traits.workspace = true
+rand.workspace = true
+serde_json.workspace = true
 serde.workspace = true
 thiserror.workspace = true
 toml.workspace = true
 tracing-subscriber = { workspace = true, features = ["env-filter"] }
 tracing.workspace = true
-serde_json.workspace = true
-figment.workspace = true
 
 [dev-dependencies]
 criterion.workspace = true
