+++ collaborative-groth16/Cargo.toml
@@ -23,16 +23,14 @@ ark-ff = { workspace = true }
 ark-groth16 = { workspace = true, optional = true }
 ark-poly = { workspace = true }
 ark-relations = { workspace = true }
-ark-serialize = { workspace = true }
-ark-std = { workspace = true }
 circom-types = { version = "0.3.0", path = "../circom-types" }
+co-circom-snarks = { version = "0.1.0", path = "../co-circom-snarks" }
 eyre = { workspace = true }
 itertools = { workspace = true }
 mpc-core = { version = "0.2.1", path = "../mpc-core" }
 mpc-net = { version = "0.1.2", path = "../mpc-net" }
 num-traits = { workspace = true }
 rand = { workspace = true }
-serde = { workspace = true }
 tracing = { workspace = true }
 
 [dev-dependencies]
