+++ release-please-config.json
@@ -24,6 +24,14 @@
       "draft": false,
       "prerelease": false
     },
+    "co-circom-snarks": {
+      "changelog-path": "CHANGELOG.md",
+      "release-type": "rust",
+      "bump-minor-pre-major": true,
+      "bump-patch-for-minor-pre-major": false,
+      "draft": false,
+      "prerelease": false
+    },
     "collaborative-circom": {
       "changelog-path": "CHANGELOG.md",
       "release-type": "rust",
